advanced 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,18 +9,18 @@
 int main(void)
 {
 	int i;
-	unsigned long long previous, current;
+	unsigned long previous, current;
 
 	previous = 1;
 	current = 2;
-	printf("%lld, %lld, ", previous, current);
+	printf("%ld, %ld, ", previous, current);
 	for (i = 0; i < 50; i++)
 	{
 		int temp = current;
 
 		current += previous;
 		previous = temp;
-		printf("%lld", current);
+		printf("%ld", current);
 		if (i != 49)
 			printf(", ");
 	}
